From Gideon May, compile fixes for OSX.

+++ src/osg/ApplicationUsage.cpp
@@ -222,7 +222,7 @@ void ApplicationUsage::writeEnvironmentSettings(std::ostream& output)
     {
         std::string::size_type len = citr->first.find_first_of("\n\r\t ");
         if (len == std::string::npos) len = citr->first.size();
-        maxNumCharsInOptions = maximum(maxNumCharsInOptions,len);
+        maxNumCharsInOptions = maximum(maxNumCharsInOptions,(unsigned int) len);
     }
     
     unsigned int optionPos = 2;
